Change the input and output port
This commit change the structure and the API of the simple ports.
Now, you first have to get the port, then use it.
Before :
inputs.send("input", 32)
Now :
let port = inputs.get("input").unwrap();
port.send(32);

This change is made for simpler integration of the array ports.
Also, the API is cleaner because everything do less,and the error
will be more easy to handle