Implement the array port
It is now possible to use array port in the component.
This is an untested version.

Each component have two type of input and output port : simple and
array.
The simple ports are HashMap<str, Sender or Receiver>.
The array ports are HashMap<str, HashMap<str, Sender or Receiver>.
So with all the features of HashMap, you can found your port.