Add the option default port
It's possible to defined an option port of a type that had the "Clone"
trait.

This port is a little special. It retains the last IP received and
send it again. If there is new IPs, only the last one is take in account
to replace the current option.

If there is no IP at all in the channel, the method option.recv() will
block until there is some.