[FIX] bug in the Scheduler
The scheduler had a bug were, sometimes, the component wasn't restart if
there are IPs in his channel.
It was due because the number of IP was updated after the scheduler was
aware that an IP was send.

There is two ways for a component to start :
1) A component send a message to the component (two steps : update the
number of IP in the channel and send a message to the scheduler)
2) The component end his run, and look if he had new IP.

That leads to several cases for restart:
- The component is halted. If (1), the component start.
- The component is running. If (1), the component doesn't start because
he his already running.
- The component is running. If (2), The component look if there is at
least one IP in one input port. If true, it restarts, otherwise not.

The bug fixed was, in chronological time :
- The component is running
- Another component send a start message to the scheduler
- The scheduler try to start thec component, but it is already running
- The component halt, there is no new IP in his channel
- The other component update the number of IPs in the channel <-- TOO LATE

Now, we have :
- The component is running
- The other component update the number of IPs in the channel <-- FIXED
- The other component send a start message to the scheduler
- The scheduler try to start thec component, but it is already running
- The component halt, there is a new IP in his channel, it restart

We see that the two last lines can be inverted. This is why the
scheduling rules work.