IIP in the subnet
It's possible to add IIP in the subnet.
They are mandatory of the type String, because it is impossible to
compile them without risk.

I believe that each component with an option port will provide the
component "StringToOptionContract". It will accept an IIP and cast it
fo the input port.