Problem: Subnets don't have a dep hierarchy
Solution:
There are two solutions to this:

1) the developer needs to explicitly describe the dependencies
of a subnet by passing in a list of derivation names to
subnetDeps. This then symlinking files into the $out/lib which ensures
those subnetDeps are registered as dependencies.

The drawback of this solution is:
A developer can put any derivation in subnetDeps and nix will
register it without throwing an error. This means that fvm
will query an incorrect MANIFEST file which, then pull down a
dependency tree that will not result in the correct execution of
the subnet.  This results in a user runtime error. Pathetic.

2) The second solution is much better, but I'm uncertain nix
will allow it:
During the buildPhase of buildFractalideSubnet, the contents of
every parentheses should be extracted and along with a nix expression
written into a /tmp/nix-build-maths_boolean_and.drv-0/lib/subnetDeps.nix
then while in the same buildPhase an ${import /path/to/subnetDeps.nix}
takes place thus importing the deps accordingly.

I highly doubt this solution will work but it would save us
from having to deal with un/intentionally incorrect dependency trees.