Problem: fvm must not test subnets.
Solution: fvm must not test subnets during the checkPhase as
it would cause an infinite recursion - or - the contract_lookup and
component_lookup wouldn't be available to fvm. Thus the boundaries
of fvm must be defined at the point whereby it can implement a
basic operating system such as Nix and Guix. The Fractalide repository
will not be all encompasing assimilating every fbp app out there.
This must be the job of another repo, written completely in fbp.