Problem: no lazy compilation
Solution: rely on nix completely which has lazy
evaluation. Create a wrapper around nix-build/nix-shell
which will download the dependencies beforehand.

process:
../fractalide $ nix-build
builds a shell script which builds a particular instance of fvm
depending on the flow-based programming file inserted.
fvm is now the generally executed file which allows the user to
create a new fvm whenever the contents of the *.fbp file changes.