Problem: cannot use crates.io rustfbp for incremental recompilation
Solution: add a --argstr local-rustfbp which allows
a fractalide programmer interested in making changes
to rustfbp code *without* having to upload it to crates.io
each time.

The flag is for the incremental compilation option coming
in which allows a stabilized rustfbp crates.io dependency
to be cached, which is needed for incremental compilation.