Problem: no default component set when building
Solution: make fsh, fractalide shell the default
build option. fsh is the main entry point to the rest
of fractalide. It abstracts away nix and building
components and merely presents a high level piping type language.
It's a reduced expression but the main goal of this level of
abstraction is dynamically creating initial information packets
from the command line.