Add a trait `OutputSend`
The goal of this trait is to help the user to play with the type
`Option<IPSender>`. This type is used by the user of the `agent`.

With this trait, the user can

```
try!(self.output.name.send(msg));
```

and must not do :

```
if let Some(sender) = self.output.name {
   try!(sender.send(msg));
} else { /* Error */ }
```

If the Option is `None`, it signify that the port is not connected.