[FIX] [rustfbp] prob: the graph doesn't stop
Until now, the graph was always running. Even if all computations were done,
there was no way to stop the graph.

This commit apport a solution. Each agent now return a `Signal` to the
`scheduler`. There are two cases :

- `Ok(End)` : Say to the scheduler that the graph can stop. The agent doesn't
  wait any more messages.
- `Ok(Continue)`: Say to the scheduler that the graph must continue to run. The
  agent wants more messages to end the computation.

An example of agent that want to use `Ok(Continue)` is the agent scheduler.
You can send to the scheduler agent "add this agent", "connect them", "send an
initial msg", ... . But the computation of the scheduler doesn't end when there
is no more input messages, it ends when the graph inside it ends. So the big
picture is :

```
recv message to modify the graph:
    manage the message;
    return Ok(Contine);
or recv halt message:
    wait for the end of the inner graph;
    return OK(End);
```

Or a http server :

```
recv request:
    send it outside;
    return Ok(Continue); // Perhaps new request will appear in the graph
recv response:
    send to the client;
    return Ok(Continue); // Perhaps new request will appear in the graph
recv halt:
    // perhaps wait the last response?
    return Ok(End);
```