[rustfbp] agent.send_action() -> send_action!()
Before, for sending an action, there was :

```
self.send_action("output", msg)?;
```

that send the msg on the correct output port.

Now, thi s is :

```
send_action!(self, output, msg)?;
```

This is a macro. This is more powerfull because the name of the output port is
tested at compile time, this cannot fail at run-time for typo.
Also, this is more efficient because there is less test/branch/instruction.