Problem: capnp approach still not ready for targetting multiple backends
Solution: this stage completes the most important step
bing the removal of complicated capnp imports and
and the removal of the capnp ids, which are now auto generated.
It also means we lose determinism, so a later problem will be to
reverse engineer capnp's id creation mechanism.

A separate derivation is used to unify the schema. This then
allows for faster compilation of the core agent.

Next step is to make available the edge.capnp for the subnet
level usage of capnp tool.