Problem: we are building racket-minimal every time
The racket with libiconv override (in racket2nix) is not available in
hydra. We don't really need the override unless we build racket-doc.

Solution: Supply an up-to-date racket-minimal from the latest nixpkgs.

This lowers test build time from ~30 minutes to ~20 minutes.

We still have the problem that it's building nix-2.0. This is probably
related to NixOS/nixpkgs#38280 .