Problem : no clean halt of the scheduler
Solution : Retain the number of running agent. Once the user want to stop (send
the msg-stop), we will wait that the number of running agent is 0, and then stop
the scheduler.

known problem :
- We add to add a sleep in the main.rkt, otherwise the scheduler receive the
msg-stop before the execution of the first agent. Not yet sure how to fix it.
- In RustFBP, a component can keep the execution running, for example if it run
an sub-thread. We will need to add this possibility.