Problem: rkt-fbp doesn't package
Because we have the info.rkt in the root of our repo, racket expects
the modules to be referred to as fractalide/modules/rkt/rkt-fbp/...

The typed-racket-more dependency is not declared.

Solution: Adjust the require{,/typed} statements to use the longer
path. Add the typed-racket-more dependency.

It's not pretty. We should probably reconsider our repo structure.