Problem: pkgs.fractalide is sensitive to result symlinks
When you nix-build, normally you will leave behind a `result` symlink
in the directory where you ran it. When our whole repo is a source
input to `pkgs.fractalide`, that means that every time you run
nix-build inside the repo, it needs to build again, even if that
symlink is all that changed.

Solution: Filter out symlinks that start with 'result'.