Problem: CHANGELOG is not machine-readable
Its function as a timeline is outdated. We have commit history.

Solution: Use `CHANGELOG.json` to capture the current state of development.

The website can parse this and show the progress in a nice and
visualized way.

To view progress over time, just `git log -p CHANGELOG.json`.

This first version of `CHANGELOG.json` is not to be taken too seriously.