Problem: No way to build dynamic GUI
Solution: provide `(dynamic-add ...)` and `(dynamic-remove ...)`

These method can be used in any card that are eventually connected to a frame.
They are used like :
```
(dynamic-add a-graph input output)
```
where `a-graph` can contains any node (agent or sub-graph). If `a-graph`
contains an virtual output port named "dynamic-out", it will be linked to the
input port "in" of the component.

`(dynamic-remove a-graph input output)` does the exact oposite, it removes the
given graph.

An example is showed in `test/dynamic`.