Problem: Action use vector and string
Solution: move for `cons` and `quote`

An action message have an "action", and a message. There is no need for more
than two field (possible with the use of struct msg).
Also, the action is only use for match, and then a `string` is a too complex and
heavy data structure, `quote` is more addapted.