Problem: node names use "/" which nix recognizes as a path
Solution:

We now use the "." to indicate a namespace which when compiling
as racket simply replaces the "." with a "/".
Using "." meant to map directly to a nix attribute name, which
further entails that we cannot name a node like "maths_boolean_nand"
instead we have to compose trees of attributes in nix, thereby
the attribute names should be "maths.boolean.nand".

This is rather annoying because at every graph level directory
we need to have a default.nix mapping each of those nodes. I'm
sure this can be automated somehow. A problem to be solved later.