Problem: no procedure to require agents by short symbol
Solution: Implement dynamic-require-agent.

If it fails with a plain dynamic-require and the agent is a symbol,
it retries within the agents directory of the fractalide collection.

This will allow us to become independent of which current directory
we are in.

Find and replace uses of dynamic-require for graphs and agents.

Yes, we have graphs in the agents directory.