Problem: No place-holder in GUI
Solution: add it

This widget allows to choice between different sub-card to display.

All the children must be in the "place" input array port. Then, once created,
they can send the action `'(display . #t)` to be show in the place-holder. There
is only one children displayed at the time, and the creation doesn't show them.

```
(node "ph" ${gui.place-holder})
(node "1" ${a-card})
(node "2" ${a-card})
(mesg "1" "in" '(init . ...))
(mesg "2" "in" '(init . ...))
(mesg "1" "in" '(display . #t))
```
will display the "1" in the place-holder, and not "2" (even if this last one is
already created).

the message
```
(mesg "2" "in" '(display . #t))
```
will display the "2" and hide "1", but this last one is not deleted.

ie : in test/main.rkt, the `choice` widget allows to display a specific counter
between three.