Problem: require an edge was not nix friendly
solution: add `(require/edge ...)`

Now, it is possible to use the macro `require/edge` in agent and graph.
It is designed as follow :
- `(require/edge ${fvm.dynamic-add}`
will look for `fractalide/modules/rkt/rkt-fbp/edges/fvm/dynamic-add`
- `(require/edge (file "/full/path/to/edge.rkt"))`
will load the full path.

So if there is no nix pre-processing, we will be in the former case.
If nix replace `${...}` by `(file "the/path)`, we will be in the second case.