Problem: No tab-panel in GUI widget
Solution: add it

It is possible to create a ${gui.tab-panel}.
The tab-panel is used only with the "places" input array port.
A selection is this port is of type "1;name". The number before the `;` is use
for ordering the tabs, the name after `;` is the name of the tabs.