Problem: No paging-jsend-get node
Solution: Create it

It should accept a request talk to an http node for the actual
fetching, parse incoming JSON, figure out the paging metadata, and
send a sequence of data objects to the output. It will be a good test
case for packaging, separation of nodes and edges, test scaffolding,
all that.

This first draft merely compiles.