Problem: nodes: No catalog.rktd
Solution: Add script for generating catalog from package. Generate it.

Also clean up dependencies.

Fractalide is not a dependency. Only the test requires it. We
don't have a good mechanism in place for that situation yet, and
we don't want a circular dependency.

The test module should load everything fractalide dynamically, so
that the package compiles without fractalide.