Problem: paging-jsend-get: Test loads paging-jsend-get
This collection no longer exists.

Solution: Load fractalide/nodes/rkt/paging-jsend-get/main instead.

`raco test .` in the root of fractalide successfully runs the test.

This should be in `release.nix` and more modules should have tests in
their `test` submodule. But first we should make running node tests
more convenient, which I guess goes under #238. Once we have that,
quicktest is just one way to formulate a test case.

The way paging-jsend-get tests itself can be viewed as a first
prototype and an example of what kind of boilerplate we would
like to put into a convenience interface.

Instead of being an external module, test/mock could be a submodule --
if we think spreading files around doesn't look nice and if the mock
is small enough. It's possible that we could make a mock generator
macro that creates a mock for your needs as a submodule, to keep the
mock definition concise.