Problem: No demux agent
We want this for being able to take e.g. a bunch of buttons and
defining their different actions in the graph, without having to loop
over an input array in the node that takes action.

Solution: Implement demux agent.

For each message Y coming on `in` selection X, message (X . Y) goes
to `out`.

Implement `identity` agent to be able to test demux[0]: Everything
that comes on `in` goes to `out`.

[0] Array port selections don't exist until you connect to them, so
    we need something to connect.