Problem: No generic agent for multiple inputs
Solution: Implement ${plumbing.transform-ins-msgs}.

A function is provided to `"option"`.

It will be called as soon as all selections on "in" have received a
message. After that it is called every time a messages comes in on one
of the "in" selections.

The function receives a hash with the selections as keys, the latest
message received on that selection as values. It returns a list of
messages to be sent.