Problem: Not detecting hung tests
Our tests are only detecting faulty signals, not whether you'll have
to wait forever for the signal.

Solution: Run each test suite with a timeout.

Also measure the times, so we can see how far we are from the limit.

Currently all of the tests run in under 10 s each on my machine. Files
with no tests discovered take 5--6 s. So my chosen timeout of 20 s has
a 3x margin on the actual test portion.