Problem: raco test doesn't work
```
/nix/store/8b2ycliakg6zsisbqcihfs9wlzgpdgsg-racket-minimal-6.12/bin/racket: Unrecognized command: test
```

The problem is that although `compiler-lib` is added as a source to
fractalide-rkt-tests, it is not filtered out as a dependency, so it's
never installed.

The filter acts on `name`, but racket2nix now names the derivation
`<racket-derivation>-<pname>`, so the filter looks for `compiler-lib`
and lets `racket-6.12-compiler-lib` through.

Solution: Filter `fractalide-rkt-tests` dependencies on `pname`.

 - Bump to fractalide/racket2nix#190 which makes `pname` available.

Closes fractalide/racket2nix#189