Problem: Building a GUI is complex
Solution: This is proposition that take another direction that the one taken
until now. I believe it will be more optimal, less boiler-plate, and easier to
thing about, but that must be tested on real program to be confirm...

Before this commit :
The UI is represented as a graph. Each node of the graph keep a state, and can
change the UI. So to have two counters, you need to have two nodes "${counter}"
in the graph.

Proposition after this commit:
See the UI as any other Message on the graph. The UI is a different
representation of data.
The main idea is :

```
(counter . (counter id value)) -> input counter() output -> input frame()
counter() output['counter] -> input display()
```
Here, the `counter()` receive a mesg of type `counter`, and generate a part of
the UI. Also, this `counter()` will generate the event linked to the model.
A more concrete example is implemented in `${guiv2.test}`.
If several counter are created in the `counter()` agent, this agent will
generate all the event for all the counters.

Advantage :
- No dynamic graph, node node that is spawned in live.
- Much less mesg in the graph. Each node can manage the event he wants.
- Seems to fit more the "data-driven" paradigm. There is much less code (no "gui
framework"). Hope there will be less friction.
- More flexibility on who do what (compute a new state, manage an event, ...)

Cons :
- The UI is no more a visual graph.
- There is much more "redraw everything". Not sure it will be easy to save all
the state.

Much more to discover...