Problem: rkt-tests: Fails because cantor and guiv2 require gui
This breaks the test runner.

Instead of adding them to the exclusion list, accept that only 10 of
our rkt files contain tests.

Solution: Test only files that declare a test submodule.

As a bonus, this drastically cuts down testing time. Racket takes a
very long time (20 s -- 30 s) just to load a file, only to find out
there are no tests. This is now done on 10 files instead of 172 files.

Grepping for 'module+ test' is not a generic way of finding tests, but
it works on our code.